rename and reorder some things
