+++ resources/js/books/UserBookLikesContext.ts
@@ -1,0 +1,6 @@
+import { createContext } from 'react';
+
+export const UserBookLikesContext = createContext({
+    userBookLikes: null,
+    setUserBookLikes: (newUserBookLikes: any) => {},
+});
