--- resources/js/books/context/UserLikesContext.ts
@@ -1,6 +1,0 @@
-import { createContext } from 'react';
-
-export const UserLikesContext = createContext({
-    userLikes: null,
-    setUserLikes: (newUserLikes: any) => {},
-});
